add bill CRUD

+++ restservice/src/main/java/com/ukrlegislation/entities/Bill.java
@@ -40,7 +40,7 @@ public class Bill implements Serializable {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
